% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_equivalence.R
\name{box_equivalence}
\alias{box_equivalence}
\title{Mapping Between Regular SSM Grid Cells and Atlantis Polygons}
\usage{
box_equivalence(
  SSM_file_path,
  SSM_file_name,
  RegularSSM_file_path,
  RegularSSM_file_name
)
}
\arguments{
\item{SSM_file_path}{Path to a SSM file. Used to load `siglev` and `siglay` information.}

\item{SSM_file_name}{Name of the SSM file (Hydrologic file).}

\item{RegularSSM_file_path}{Path to the SSM file with a regular grid (after interpolation).}

\item{RegularSSM_file_name}{Regular grid SSM file containing TS (temperature/salinity) variables.}
}
\value{
No value is returned. One CSV file is written in `R/code`:
`box_composition.csv`.
}
\description{
Creates the correspondence between regular SSM grid cells and Atlantis polygons.
A single CSV file is written to the `R/code` directory:
`box_composition.csv`.
If this file already exists, you can skip to Step 1.
}
\examples{
SSM_file_path        <- "/nfsdata/SSM_example/HYD"
SSM_file_name        <- "ssm_00117.nc"
RegularSSM_file_path <- "here(File_regular_grid/bau/2011)"
RegularSSM_file_name <- "regular_grid_TS_2011.nc"
box_equivalence(SSM_file_path, SSM_file_name,
                            RegularSSM_file_path, RegularSSM_file_name)

}
