% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_horizontal_equivalence.R
\name{face_horizontal_equivalence}
\alias{face_horizontal_equivalence}
\title{Mapping Between Regular SSM Horizontal Faces and Atlantis Polygons}
\usage{
face_horizontal_equivalence(
  SSM_file_path,
  SSM_file_name,
  RegularSSM_file_path,
  RegularSSM_file_name
)
}
\arguments{
\item{SSM_file_path}{Path to a SSM file. It is used to load the siglev and siglay information}

\item{SSM_file_name}{SSM file name (Hydrologic file)}

\item{RegularSSM_file_path}{Path to a SSM file with regular grid after interpolation}

\item{RegularSSM_file_name}{Regular grid SSM file (must contain variable `W`)}
}
\value{
No value is returned. Two CSV files are written in R/code:
`box_composition_ww.csv` and `layer_max_ww.csv`
}
\description{
Creates the correspondence between regular grid SSM cells and Atlantis polygons for horizontal faces
(i.e., the bottom of each polygon). Two CSV files are written in R/code:
`box_composition_ww.csv` and `layer_max_ww.csv`.
The sign of `w` is positive downwards (downwelling), negative upwards (upwelling).
If these files already exist, you can skip to Step 2.
}
\examples{
SSM_file_path        <- "/nfsdata/SSM_example/HYD"
SSM_file_name        <- "ssm_00117.nc"
RegularSSM_file_path <- "~/UWatlantisgroupvm05/amps_hydrodynamics/Workflow/Step A/File_regular_grid/No_VELMA/2011"
RegularSSM_file_name <- "regular_grid_UVW_novelma_2011.nc"
face_horizontal_equivalence(SSM_file_path, SSM_file_name,
                            RegularSSM_file_path, RegularSSM_file_name)

}
