% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_vertical_equivalence.R
\name{face_vertical_equivalence}
\alias{face_vertical_equivalence}
\title{Mapping Between Regular SSM Vertical Faces and Atlantis Polygons}
\usage{
face_vertical_equivalence(
  SSM_file_path,
  SSM_file_name,
  RegularSSM_file_path,
  RegularSSM_file_name
)
}
\arguments{
\item{SSM_file_path}{Path to a SSM file. It is used to load the `siglev` and `siglay` information.}

\item{SSM_file_name}{SSM file name (Hydrologic file).}

\item{RegularSSM_file_path}{Path to a SSM file with regular grid after interpolation.}

\item{RegularSSM_file_name}{Regular grid SSM file containing UV variables.}
}
\value{
No value is returned. Two CSV files are written in `R/code`:
`uxy2_uv_code.csv` and `face_composition_uv.csv`.
}
\description{
Creates the correspondence between regular grid SSM cells and Atlantis polygons for vertical faces
(i.e., the borders between polygons). Two CSV files are written in `R/code`:
`uxy2_uv_code.csv` and `face_composition_uv.csv`.
If these files already exist, you can skip to Step 3.
}
\examples{
SSM_file_path        <- "/nfsdata/SSM_example/HYD"
SSM_file_name        <- "ssm_00117.nc"
RegularSSM_file_path <- "here(File_regular_grid/bau/2011)"
RegularSSM_file_name <- "regular_grid_UVW_2011.nc"
face_vertical_equivalence(SSM_file_path, SSM_file_name,
                          RegularSSM_file_path, RegularSSM_file_name)

}
