% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dailyfiles.R
\name{join_dailyfiles}
\alias{join_dailyfiles}
\title{join_dailyfiles: Merge the files created in step B into one file}
\usage{
join_dailyfiles(year, variable, scenario)
}
\arguments{
\item{year}{Year to process.}

\item{variable}{Variable name (e.g., "temperature", "NO3").}

\item{scenario}{Scenario name (e.g., "Hist", "RCP85").}
}
\value{
Nothing. Side effects: saves merged file.
}
\description{
join_dailyfiles: Merge the files created in step B into one file
}
\examples{
join_dailyfiles(2011, "temperature", "BAU")
}
