% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing Check and Environment Initialization}
\usage{
preprocess()
}
\value{
No return value. Messages are printed to the console.
}
\description{
Before using the main functions of the SSMtoAtlantis package, this helper function:
- Ensures that the Python environment `Salish_Sea_env` is activated (via R script)
- Verifies that key preprocessing files are present
- Guides the user to run the necessary equivalence functions if any files are missing
}
