% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python_env.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Create and configure a conda environment for the SSM to Atlantis translation}
\usage{
setup_python_env(env_name = "Salish_sea_env5", conda_path)
}
\arguments{
\item{env_name}{Character. Name of the conda environment to create or use. Default is `"Salish_sea_env"`.}

\item{conda_path}{Character. Full path to the `bin` directory in the anaconda or miniconda folder
used for your conda installation.}
}
\description{
This function sets up a Python conda environment for the SSM to Atlantis translation
developped in the SSMtoAtlantis package.
It adds the specified `conda_path` (path to bin the de anaconda or miniconda folder)
to the system `PATH`. Then, it checks if the conda environment `env_name`
already exists, and if not, creates it using a list of required packages from a file.
It also configures reticulate to use the new environment.
}
\examples{
setup_python_env(env_name = "Salish_sea_env",
  conda_path = "/home/atlantis/anaconda3/bin")
}
