% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMtoAtlantis.R
\name{SSMtoAtlantis}
\alias{SSMtoAtlantis}
\title{SSMtoAtlantis: Function that creates Atlantis inputs from SSM outputs (for all
the variables and in the correct Atlantis grid). It expect yearly files.}
\usage{
SSMtoAtlantis(year, scenario, filename.hyd, filename.wq)
}
\arguments{
\item{year}{The year of the SSM outputs (numeric or string).}

\item{scenario}{The name of the scenario used to produced SSM output}

\item{filename.hyd}{The input filename of the file with physical variables.
It is a NetCDF file which include SSM outputs on FVCOM grid (variables are be
currents (U,V and W) temperature or salinity).}

\item{filename.wq}{The input filename of the file with water quality variables.
It is a NetCDF file which include SSM outputs on FVCOM grid (variables are be
nutrients (NO3, NH4), phytoplankton (B1, B2), zooplankton (SZ, LZ)
organic matters (LPON, LDON, RPON, RDON)))}
}
\value{
Nothing. Creates Atlantis inputs files for the all variables.
}
\description{
SSMtoAtlantis: Function that creates Atlantis inputs from SSM outputs (for all
the variables and in the correct Atlantis grid). It expect yearly files.
}
\examples{
SSMtoAtlantis(2011, "status_quo", filename.hyd = "SSM2011_status_quo_hyd.nc",
              filename.wq = "SSM2011_status_quo_wqm.nc")
}
