% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step_A.R
\name{StepA}
\alias{StepA}
\title{StepA: From FVCOM grid to regular (ROMS like) grid for SSM outputs}
\usage{
StepA(year, variable, scenario, filename)
}
\arguments{
\item{year}{The year of the SSM outputs (numeric or string).}

\item{variable}{The variable name to extract.}

\item{scenario}{The name of the scenario used to produced SSM output}

\item{filename}{The input filename, a NetCDF file which include SSM outputs on FVCOM grid.}
}
\value{
Path to the generated NetCDF file with SSM regular output for variable (invisible).
}
\description{
StepA: From FVCOM grid to regular (ROMS like) grid for SSM outputs
}
\examples{
year = 2095
variable = "temperature"
scenario = "BAU"
filename = "path/to/SSM_FVCOM_grid_file"
StepA(year, variable, scenario, filename)
}
