% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.SSMtoAtlantis.R
\name{var.SSMtoAtlantis}
\alias{var.SSMtoAtlantis}
\title{Title}
\usage{
var.SSMtoAtlantis(year, variable, scenario, filename)
}
\arguments{
\item{year}{The year of the SSM outputs (numeric or string).}

\item{variable}{The variable name to extract.}

\item{scenario}{The name of the scenario used to produced SSM output}

\item{filename}{The input filename, a NetCDF file which include SSM outputs on FVCOM grid.}
}
\value{
Nothing. Creates Atlantis inputs files for one variable.
}
\description{
Title
}
\examples{
var.SSMtoAtlantis(2011, "temperature", "status_quo", filename = "SSM2011_status_quo_hyd.nc")
}
