% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_interpoled_data.R
\name{write_interpoled_data}
\alias{write_interpoled_data}
\title{write_interpoled_data}
\usage{
write_interpoled_data(data, start_year, end_year, variable, scenario)
}
\arguments{
\item{data}{A 3D numeric array with dimensions (or a list of 3 D numeric array with dimensions for uvw)
`layer × box × time`, containing interpolated values
for the selected variable.}

\item{start_year}{Numeric. First year of the forcing period.
Used to define the NetCDF time origin.}

\item{end_year}{Numeric. Last year of the forcing period.
Used in the output file name.}

\item{variable}{Character string. Name of the variable to write.
Must be one of:
`"salinity"`, `"temperature"`, `"U"`, `"V"`, `"W"`,
`"NO3"`, `"NH4"`, `"SZ"`, `"LZ"`, `"MZ"`, `"SP"`, `"LP"`,
`"O2"`, `"LPON"`, `"RPON"`, `"RDON"`.}

\item{scenario}{Character string. Scenario name used to define
the output directory structure.}
}
\value{
Nothing returned. .
}
\description{
Write 3D array after interpolation between two or more SSM year into
an Atlantis input NetCDF file. In a case where variable = uvw, it write a
list of 3 3D array: exchange, dest_b, dest_k
}
\details{
This function writes an interpolated 3D array (z × box × time)
into a NetCDF file formatted for the Atlantis model forcing files.
The output file is created in the `Atlantis_inputs/<scenario>/interannual`
directory and follows the naming convention used for the
Puget Sound Atlantis configuration.

The variable name, long name, and units are automatically mapped
to the corresponding Atlantis conventions.


The function:
\itemize{
  \item Creates vertical (`z`), box (`b`), and time (`t`) dimensions
  compatible with Atlantis.
  \item Assigns variable metadata (long name, units) based on
  internal correspondence with Atlantis variable names.
  \item Writes valid minimum and maximum attributes.
  \item Adds standard global attributes required for Atlantis forcing.
}

The time dimension is defined in seconds since
`start_year-01-01`, with a forced default time step (`dt`)
of 43200 seconds (12 hours).
}
\examples{
\dontrun{
# Example with a dummy array
dummy_array <- data(runif(10 * 5 * 20),
                     dim = c(10, 5, 20))

write_interpoled_data(
  data = dummy_array,
  start_year = 2011,
  end_year = 2045,
  variable = "temperature",
  scenario = "status_quo"
)
}
}
