% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_interpolation.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{linear_interpolation}
\usage{
linear_interpolation(year_start, year_end, scenario, variable)
}
\arguments{
\item{year_start}{numeric Start year}

\item{year_end}{numeric End year}

\item{scenario}{character Scenario name}

\item{variable}{character Variable name}
}
\value{
Array of interpolated values
}
\description{
linear_interpolation
}
\examples{
year_start = 2015
year_end = 2025
scenario = "status_quo"
variable = "temperature"
linear_interpolation(year_start, year_end, scenario, variable)
}
