% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepB.R
\name{StepB}
\alias{StepB}
\title{StepB: From regular (ROMS like) grid to Atlantis intputs}
\usage{
StepB(year, variable, scenario, PCB_congener = "0")
}
\arguments{
\item{year}{The year of the SSM outputs (numeric or string).}

\item{variable}{The variable name to extract. Choice are:
salinity, temperature, U, V, W, NO3, NH4, SZ, LZ, MZ, SP, LP, Oxygen, LPON, RPON, RDON, PCB}

\item{scenario}{The name of the scenario used to produced SSM output}

\item{PCB_congener}{When variable = PCB, PCB_congener informs the PCB congener number.}
}
\value{
Nothing. Creates daily variable files.
}
\description{
StepB: From regular (ROMS like) grid to Atlantis intputs
}
\examples{
StepB(2011, "temperature", "status_quo")

}
